+++ WebAuthN/Util/tpm2-tss/util/io.h
@@ -30,21 +30,21 @@ typedef SSIZE_T ssize_t;
 #define TEMP_RETRY(dest, exp) \
 {   int __ret; \
     do { \
-        __ret = exp; \
+        __ret = (int) exp; \
     } while (__ret == SOCKET_ERROR && WSAGetLastError() == WSAEINTR); \
     dest = __ret; }
 #elif defined (__FreeBSD__)
 #define TEMP_RETRY(dest, exp) \
 {   int __ret; \
     do { \
-        __ret = exp; \
+        __ret = (int) exp; \
     } while ((__ret == SOCKET_ERROR) && (errno == EINTR || errno == EAGAIN)); \
     dest =__ret; }
 #else
 #define TEMP_RETRY(dest, exp) \
 {   int __ret; \
     do { \
-        __ret = exp; \
+        __ret = (int) exp; \
     } while (__ret == SOCKET_ERROR && errno == EINTR); \
     dest =__ret; }
 #endif
